Factor out container classes as they are no longer required.
