+++ rhino/src/main/java/org/mozilla/javascript/LockAwareSlotMap.java
@@ -1,10 +1,9 @@
 package org.mozilla.javascript;
 
 /**
- * Extends {@link SlotMap} with a set of "WithLock" methods. which
- * will not acquire the lock. These should only be used internally by
- * implementation, or by other {@link SlotMap}s which share the same
- * lock.
+ * Extends {@link SlotMap} with a set of "WithLock" methods. which will not acquire the lock. These
+ * should only be used internally by implementation, or by other {@link SlotMap}s which share the
+ * same lock.
  */
 interface LockAwareSlotMap extends SlotMap {
     /** The equivalent of {@link SlotMap#size()}. */
@@ -19,7 +18,10 @@ interface LockAwareSlotMap extends SlotMap {
     /** The equivalent of {@link SlotMap#query(Object, int)}. */
     Slot queryWithLock(Object key, int index);
 
-    /** The equivalent of {@link SlotMap#compute(SlotMapOwner, Object, int, org.mozilla.javascript.SlotMap.SlotComputer)}. */
+    /**
+     * The equivalent of {@link SlotMap#compute(SlotMapOwner, Object, int,
+     * org.mozilla.javascript.SlotMap.SlotComputer)}.
+     */
     <S extends Slot> S computeWithLock(
             SlotMapOwner owner, Object key, int index, SlotComputer<S> compute);
